+++ manifest.json
@@ -14,19 +14,15 @@
       {
         "css": [
           "/src/styles.css",
-          "/src/ext/micromodal.css",
-          "/src/randomizer/random.css"
+          "/src/ext/micromodal.css"
         ],
         "js": [
           "/src/ext/micromodal.min.js",
           "/src/group-manager.js",
           "/src/bi-logger.js",
-          "/src/pimp-my-wolt.js",
-          "/src/randomizer/category-random.js"
+          "/src/pimp-my-wolt.js"
         ],
-        "matches": [
-          "https://wolt.com/*"
-        ]
+        "matches": ["https://wolt.com/*"]
       },
       {
         "css": [
@@ -50,7 +46,9 @@
           "/assets/loader.gif",
           "/assets/icons/pimp-my-wolt-icon-48.png",
           "/assets/icons/pimp-my-wolt-icon-128.png",
-          "/assets/hungry_wheel.png"
+          "/assets/hungry_wheel.png",
+          "src/wheel/wheelstyle.css",
+          "src/wheel/wheel-index.js"
         ],
         "matches": [
           "<all_urls>"
@@ -72,7 +70,6 @@
         "24": "/assets/icons/pimp-my-wolt-icon-24.png",
         "48": "/assets/icons/pimp-my-wolt-icon-48.png"
       },
-      "default_title": "Pimp My Wolt",
-      "default_popup": "popup.html"
+      "default_title": "Pimp My Wolt"
     }
   }
